Add Authentication

+++ backend/controllers/goalcontroller.js
@@ -1,16 +1,18 @@
 //when we use mongoose to interact with the database for these fucntions, we get back a promise
 const asyncHandler = require("express-async-handler");
-
 const Goal = require("../models/goalmodels");
-// Good practice for description
+const User = require('../models/usermodels')
 
+// Good practice for description
 // @desc    Get goals
 // @route:  GET /api/goals
 // @access: Private
 const getGoals = asyncHandler( async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({
+    user: req.user.id,
+  })
   //status is optional
-  res.status(200).json(goals);
+  res.status(200).json(goals); 
 })
 
 // @desc    Set goals
@@ -23,8 +25,9 @@ const setGoal = asyncHandler( async (req, res) => {
   }
   
   const goal = await Goal.create({
-      text: req.body.text}
-  );
+      text: req.body.text,
+      user: req.user.id,
+  });
 
   //status is optional
   res.status(200).json( goal );
@@ -38,10 +41,22 @@ const updateGoal = asyncHandler( async (req, res) => {
   //status is optional
   // (new mongoose.Types.ObjectId(author))
   const goal = await Goal.findById(req.params.id)
-
   if (!goal){
     res.status(400)
     throw new Error("Goal not found")
+  }
+
+  const user = await User.findById(req.user.id)
+
+  if (!user){
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  //make sure login user matches to goal user
+  if (goal.user.toString()!= user.id){
+    res.status(401)
+    throw new Error("User not authorised")
   }
 
   const updatedgoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
@@ -58,6 +73,19 @@ const deleteGoal = asyncHandler( async (req, res) => {
   if (!goal){
     res.status(400)
     throw new Error("Goal not found")
+  }
+
+  const user = await User.findById(req.user.id)
+
+  if (!user){
+    res.status(401)
+    throw new Error("User not found")
+  }
+
+  //make sure login user matches to goal user
+  if (goal.user.toString()!= user.id){
+    res.status(401)
+    throw new Error("User not authorised")
   }
 
   const deletedgoal = await Goal.findByIdAndDelete(req.params.id)
